Remove old code requiring Twisted
