+++ binance/__init__.py
@@ -4,9 +4,8 @@
 
 """
 
-__version__ = '0.7.1-async'
+__version__ = '0.7.2-async'
 
-from binance.client import Client  # noqa
-from binance.client_async import AsyncClient # noqa
+from binance.client import Client, AsyncClient  # noqa
 from binance.depthcache import DepthCacheManager  # noqa
 from binance.websockets import BinanceSocketManager  # noqa
