+++ requirements.txt
@@ -1,5 +1,4 @@
 aiohttp==2.3.10
-autobahn==18.7.1
 certifi==2018.4.16
 chardet==3.0.4
 cryptography==2.3
@@ -7,5 +6,4 @@ dateparser==0.7.0
 pyOpenSSL==19.0.0
 requests==2.21.0
 service-identity==17.0.0
-Twisted==18.7.0
 websockets==3.4
